+++ mne/stats/__init__.py
@@ -1,7 +1,8 @@
 """Functions for statistical analysis."""
 
-from .parametric import f_threshold_mway_rm, f_mway_rm, f_oneway
-from .permutations import permutation_t_test
+from .parametric import (f_threshold_mway_rm, f_mway_rm, f_oneway,
+                         _parametric_ci)
+from .permutations import permutation_t_test, _ci, _bootstrap_ci
 from .cluster_level import (permutation_cluster_test,
                             permutation_cluster_1samp_test,
                             spatio_temporal_cluster_1samp_test,
