+++ mne/utils.py
@@ -2671,45 +2671,3 @@ def open_docs(kind=None, version=None):
         raise ValueError('version must be one of %s, got %s'
                          % (version, versions))
     webbrowser.open_new_tab('https://martinos.org/mne/%s/%s' % (version, kind))
-
-
-def _parametric_ci(arr, ci=.95):
-    """Calculate the `ci`% parametric confidence interval for `arr`."""
-    from scipy import stats
-    mean, sigma = arr.mean(0), stats.sem(arr, 0)
-    # This is highly convoluted to support 17th century Scipy
-    # XXX Fix when Scipy 0.12 support is dropped!
-    # then it becomes just:
-    # return stats.t.interval(ci, loc=mean, scale=sigma, df=arr.shape[0])
-    return np.asarray([stats.t.interval(ci, arr.shape[0],
-                       loc=mean_, scale=sigma_)
-                       for mean_, sigma_ in zip(mean, sigma)]).T
-
-
-def _bootstrap_ci(arr, ci=.95, n_bootstraps=2000, statfun='mean',
-                  random_state=None):
-    """Get confidence intervals from non-parametric bootstrap."""
-    if statfun == "mean":
-        statfun = lambda x: x.mean(axis=0)  # noqa
-    elif statfun == 'median':
-        statfun = lambda x: np.median(x, axis=0)  # noqa
-    elif not callable(statfun):
-        raise ValueError("statfun must be 'mean', 'median' or callable.")
-    n_trials = arr.shape[0]
-    indices = np.arange(n_trials, dtype=int)  # BCA would be cool to have too
-    rng = check_random_state(random_state)
-    boot_indices = rng.choice(indices, replace=True,
-                              size=(n_trials, len(indices)))
-    stat = np.array([statfun(arr[inds]) for inds in boot_indices])
-    ci = (((1 - ci) / 2) * 100, ((1 - ((1 - ci) / 2))) * 100)
-    ci_low, ci_up = np.percentile(stat, ci, axis=0)
-    return np.array([ci_low, ci_up])
-
-
-def _ci(arr, ci=.95, method="bootstrap", n_bootstraps=2000, random_state=None):
-    """Calculate confidence interval. Aux function for plot_compare_evokeds."""
-    if method == "bootstrap":
-        return _bootstrap_ci(arr, ci=ci, n_bootstraps=n_bootstraps,
-                             random_state=random_state)
-    else:
-        return _parametric_ci(arr, ci=ci)
